Refactored the code to work with the new changes to the development branch. It appears that the results table is currently reading the Redux state variable correctly. Once there is data in the db, we can test that it is parsing the data properly
Relates to issues #74 and #77
