+++ client/package.json
@@ -6,10 +6,11 @@
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-places-autocomplete": "^7.2.0",
+    "react-redux": "^5.1.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.0.5",
-    "react-redux": "^5.1.0",
-    "redux": "^4.0.1"
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
