+++ client/src/App.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react';
-//import { connect } from 'react-redux';
 import './styles/ResultsTable.css';
 import './styles/FilterMenu.css';
 import Navbar from './views/Navbar';
 import ResultData from './components/ResultData';
 import FilterMenu from './components/FilterMenu';
 import Routes from './config/Routes';
-import ResultsTableContainer from '.containers/ResultsTableContainer'
 
 class App extends Component {
 
@@ -29,12 +27,6 @@ class App extends Component {
         }));
     };
 
-    // This function will be used in the future to populate the data field in state to hold
-    // the data returned from search queries
-    // __populateResultsTableData() {
-
-    // }
-
     callBackendAPI = async() => {
         const response = await fetch('/express_backend');
         const body = await response.json();
@@ -47,19 +39,18 @@ class App extends Component {
     }
 
     render() {
-        return (
-            <div className="App">
-                <Navbar />
-                {this.state.isFilterMenuVisible ?
-                    <button id="filterButton" className="show clickable" onClick={this.toggleFilterVisibility}/> :
-                    <button id="filterButton" className="hide clickable" onClick={this.toggleFilterVisibility}>Filters</button>}
-                <FilterMenu visibility={this.state.isFilterMenuVisible} />
-                {/*<ResultsTable results={this.state.data} />*/}
-                <ResultsTableContainer />
-                <Routes/>
-            </div>
-        );
+      return (
+        <div className="App">
+          <Navbar />
+          {this.state.isFilterMenuVisible ?
+            <button id="filterButton" className="show clickable" onClick={this.toggleFilterVisibility}/> :
+            <button id="filterButton" className="hide clickable" onClick={this.toggleFilterVisibility}>Filters</button>}
+          <FilterMenu visibility={this.state.isFilterMenuVisible} />
+          {/*<ResultsTable results={this.state.data} />*/}
+          <Routes/>
+        </div>
+      );
     }
 }
-export default App;
-//export default connect(mapStateToProps, mapDispatchToProps)(App);
+
+export default App
