+++ client/src/reducers/rootReducer.js
@@ -1,6 +1,7 @@
 
 import { combineReducers } from 'redux';
 import postingsReducer from './postingsReducer.js';
+
 export default combineReducers({
  postingsReducer
 });
