+++ client/src/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers/RootReducer';
+import rootReducer from './reducers/rootReducer';
 
 export default function configureStore(initialState={}) {
  return createStore(
